Changed Entity Id from integer to long. Added Account email verification feature after registration through sending an email. Added some service and rest functions

+++ src/main/java/cz/cvut/fel/poustka/daniel/flashcards_backend/dao/genericDao/BaseDao.java
@@ -24,7 +24,7 @@ public abstract class BaseDao<T> implements GenericDao<T>, SpecialDao<T>
     }
 
     @Override
-    public T find(Integer id)
+    public T find(Long id)
     {
         Objects.requireNonNull(id);
         return em.find(type, id);
@@ -108,7 +108,7 @@ public abstract class BaseDao<T> implements GenericDao<T>, SpecialDao<T>
     }
 
     @Override
-    public boolean exists(Integer id)
+    public boolean exists(Long id)
     {
         return id != null && em.find(type, id) != null;
     }
